pioneer/findtopic/requirements.txt
django==2.2.7
djangorestframework==3.10.3
django-rest-swagger==2.2.0
mysqlclient==1.4.6
